# docs/元数据服务/openapi.json
{
  "components": {
    "schemas": {
...
      "_listRequest": {
        "properties": {
          "query": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "typeName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "classification": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "limit": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "offse": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          }
        }
      },
...
      "_entity_detailRequest": {
        "properties": {
          "typeName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "dbName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "tableName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "server": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "_listResponse": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "properties": {
                "guid": {
                  "format": "string",
                  "description": "数据唯一标识",
                  "type": "string"
                },
                "typeName": {
                  "format": "string",
                  "description": "元数据类别 如mysql_table mysql_db hive_table等",
                  "type": "string"
                },
                "displayName": {
                  "format": "string",
                  "description": "展示的名字",
                  "type": "string"
                },
                "qualifiedName": {
                  "format": "string",
                  "description": "限定名称",
                  "type": "string"
                },
                "createTime": {
                  "format": "string",
                  "description": "创建时间",
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "_entity_detailResponse": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "entities": {
                "description": "No comments found.",
                "type": "array",
                "items": {
                  "properties": {
                    "guid": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "displayText": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "typeName": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    }
                  }
                }
              },
              "columns": {
                "description": "No comments found.",
                "type": "array",
                "items": {
                  "properties": {
                    "owner": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "name": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "type": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "comment": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    }
                  }
                }
              },
              "tables": {
                "description": "No comments found.",
                "type": "array",
                "items": {
                  "properties": {
                    "guid": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "typeName": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "entityStatus": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "displayText": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "relationshipType": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "relationshipGuid": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    },
                    "relationshipStatus": {
                      "format": "string",
                      "description": "No comments found.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://172.20.0.13:8782/mtd"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/entity/detail": {
      "post": {
        "summary": "获取元数据详情信息",
        "deprecated": false,
        "description": "获取元数据详情信息",
        "tags": [
          "MtdApiController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_entity_detailRequest"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"typeName\": \"rochel.emard\",\n  \"dbName\": \"rochel.emard\",\n  \"tableName\": \"rochel.emard\",\n  \"server\": \"tihkmj\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_entity_detailResponse"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"y15xh6\",\n  \"tips\": \"5v1w09\",\n  \"data\": {\n    \"entities\": [\n      {\n        \"guid\": \"59\",\n        \"displayText\": \"t0jubf\",\n        \"typeName\": \"rochel.emard\"\n      }\n    ],\n    \"columns\": [\n      {\n        \"owner\": \"e9qm0d\",\n        \"name\": \"rochel.emard\",\n        \"type\": \"h7mlv0\",\n        \"comment\": \"gxffcb\"\n      }\n    ],\n    \"tables\": [\n      {\n        \"guid\": \"59\",\n        \"typeName\": \"rochel.emard\",\n        \"entityStatus\": \"of8zz7\",\n        \"displayText\": \"8i2d5d\",\n        \"relationshipType\": \"h31eou\",\n        \"relationshipGuid\": \"59\",\n        \"relationshipStatus\": \"qzjyu8\"\n      }\n    ]\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "mtdDetail",
        "parameters": []
      }
    },
...
    "/list": {
      "post": {
        "summary": "查询元数据列表",
        "deprecated": false,
        "description": "查询元数据列表",
        "tags": [
          "MtdAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_listRequest"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"query\": \"82fmjg\",\n  \"typeName\": \"rochel.emard\",\n  \"classification\": \"uz43c0\",\n  \"limit\": 10,\n  \"offse\": 932\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_listResponse"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"yvutno\",\n  \"tips\": \"ojg97y\",\n  \"data\": [\n    {\n      \"guid\": \"59\",\n      \"typeName\": \"rochel.emard\",\n      \"displayName\": \"rochel.emard\",\n      \"qualifiedName\": \"rochel.emard\",\n      \"createTime\": \"2021-10-25 16:17:38\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "searchList",
        "parameters": []
      }
    },
...
  },
  "info": {
    "title": "dm-metadata",
    "version": "1.0.0"
  }
}